+++ packages/server/src/api/utils/request.ts
@@ -37,7 +37,9 @@ export default async function request(
     );
   }
 
-  updatedHeaders.set('host', requestUrl.host);
+  if (config.skipHostHeader !== true) {
+    updatedHeaders.set('host', requestUrl.host);
+  }
 
   if (config.callbackUrl) {
     const cbUrl = new URL(config.callbackUrl);
@@ -81,8 +83,17 @@ export default async function request(
       updatedHeaders.set('content-type', 'application/json');
 
       const bodyStream = _init.body ?? _init.request?.body ?? request.body;
+      let bodyText: string;
 
-      const bodyText = await new Response(bodyStream).text();
+      if (bodyStream === request.body) {
+        try {
+          bodyText = await request.clone().text();
+        } catch (e) {
+          bodyText = await new Response(bodyStream).text();
+        }
+      } else {
+        bodyText = await new Response(bodyStream).text();
+      }
 
       // try to parse JSON, fallback to text if not
       try {
@@ -99,6 +110,7 @@ export default async function request(
   const fullUrl = `${url}${requestUrl.search}`;
 
   if (config.debug) {
+    // something going on with `fetch` in nextjs, possibly other places
     // something going on with `fetch` in nextjs, possibly other places
     // hot-reloading does not always give back `set-cookie` from fetchCSRF
     // cURL seems to always do it (and in a real app, you don't have hot reloading),